fix(reader): reliable bar show/hide + 44pt tap targets
- Taps now carry their vertical position: with bars visible, only a
  center-zone tap (22-78%) hides them; taps near the top/bottom bars are
  ignored, so narrowly missing a bar button no longer dismisses the bar
- Any tap while bars are hidden brings them back; collapsed pills are now
  tappable with a 120x44 invisible hit area
- Reader icon buttons and page-step buttons get HIG-minimum 44pt hit areas
  (visuals unchanged)
- Page navigation resets the auto-hide timer so the bar doesn't vanish
  mid-interaction

+++ PolyglotReader/Views/PDF/PDFKitView.swift
@@ -24,7 +24,9 @@ struct PDFKitView: UIViewRepresentable {
     var onSelection: ((String, CGRect, Int, [CGRect]) -> Void)?  // pdfRects added for PDF coordinate system
     var onImageSelection: ((PDFImageInfo) -> Void)?
     var onRenderComplete: (() -> Void)?
-    var onTap: (() -> Void)?
+    /// Boş alana tek dokunuş. Parametre, dokunuşun görünümdeki dikey konumu
+    /// (0 = üst, 1 = alt) — okuyucu bar toggle'ının bölge mantığı için.
+    var onTap: ((CGFloat) -> Void)?
     var onAnnotationTap: ((Annotation) -> Void)?
     /// İki parmak çift dokunuş — okuyucunun odak modu geçişini tetikler.
     var onTwoFingerDoubleTap: (() -> Void)?
